+++ CountlyContentConfig.h
@@ -8,7 +8,7 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
-#if (TARGET_OS_IOS)
+#if (TARGET_OS_IOS || TARGET_OS_VISION)
 typedef enum : NSUInteger
 {
     COMPLETED,
@@ -32,7 +32,7 @@ typedef BOOL (^ContentURLHandler)(NSURL *url);
 
 @interface CountlyContentConfig : NSObject
 
-#if (TARGET_OS_IOS)
+#if (TARGET_OS_IOS || TARGET_OS_VISION)
 /**
  * This is an experimental feature and it can have breaking changes
  * Register global completion blocks to be executed on content.