+++ CountlyWebViewController.h
@@ -4,14 +4,14 @@
 //
 // Please visit www.count.ly for more information.
 
-#if (TARGET_OS_IOS)
+#if (TARGET_OS_IOS || TARGET_OS_VISION)
   #import <UIKit/UIKit.h>
 #endif
 
 #import "CountlyCommon.h"
 
 NS_ASSUME_NONNULL_BEGIN
-#if (TARGET_OS_IOS)
+#if (TARGET_OS_IOS || TARGET_OS_VISION)
 @interface                           CountlyWebViewController : UIViewController
 @property(nonatomic, strong) UIView *contentView;
 - (void)updatePlacementRespectToSafeAreas;