+++ PassThroughBackgroundView.h
@@ -5,7 +5,7 @@
 // Please visit www.count.ly for more information.
 
 
-#if (TARGET_OS_IOS)
+#if (TARGET_OS_IOS || TARGET_OS_VISION)
 #import <UIKit/UIKit.h>
 #import <WebKit/WebKit.h>
 #endif
@@ -13,7 +13,7 @@
 #import "CountlyCommon.h"
 
 NS_ASSUME_NONNULL_BEGIN
-#if (TARGET_OS_IOS)
+#if (TARGET_OS_IOS || TARGET_OS_VISION)
 @interface PassThroughBackgroundView : UIView